+++ routes/api/blog_routes.js
@@ -9,8 +9,8 @@ import {
   findBlog,
   deleteBlog,
   likeBlog,
-  commentToBlog,
 } from "../../controllers/blogsController.js";
+import { commentToBlog } from "../../controllers/commentsController.js";
 import multer from "multer";
 import verifyUserToken from "../../middleware/authVerifyMiddleWare.js";
 
@@ -27,6 +27,12 @@ router
   );
 
 router
+  .route("/:id/comments")
+  .post([verifyUserToken, validate(comment_schema)], commentToBlog);
+
+router.route("/:id/likes").get(verifyUserToken, likeBlog);
+
+router
   .route("/:id")
   .put(
     [verifyUserToken, upload.single("image"), validate(blog_schema)],
@@ -34,11 +40,5 @@ router
   )
   .get(findBlog)
   .delete(verifyUserToken, deleteBlog);
-
-router.route("/like/:id").get(verifyUserToken, likeBlog);
-
-router
-  .route("/comment/:id")
-  .post([verifyUserToken, validate(comment_schema)], commentToBlog);
 
 export default router;
